+++ packages/plugin-inquirer-cli/test/src/Cli.spec.js
@@ -257,7 +257,7 @@ describe("Cli", () => {
 
     it("should refresh main menu", async () => {
       expect.assertions(2);
-      coreInstance.alerts.onChange.getCall(0).args[0]();
+      coreInstance.alerts.root.onChange.getCall(0).args[0]();
       expect(inquirerMocks.stubs.inquirer.inquire.callCount).toEqual(2);
       expect(inquirerMocks.stubs.inquirer.inquire.getCall(1).args[0]).toEqual("main");
     });
@@ -307,7 +307,7 @@ describe("Cli", () => {
     beforeEach(async () => {
       removeChangeMocksSpy = sinon.spy();
       removeChangeAlertsSpy = sinon.spy();
-      coreInstance.alerts.onChange.returns(removeChangeAlertsSpy);
+      coreInstance.alerts.root.onChange.returns(removeChangeAlertsSpy);
       coreInstance.mock.onChange.returns(removeChangeMocksSpy);
       await cli.start();
     });